Classifier --> QueryIndexer, KmerExtractor, KmerMatcher, Taxonomer, Reporter
