+++ src/commons/common.h
@@ -6,6 +6,7 @@
 
 #define likely(x) __builtin_expect((x),1)
 #define unlikely(x) __builtin_expect((x),0)
+#define kmerLength 8
 
 struct SequenceBlock{
     SequenceBlock(size_t start, size_t end, size_t length, size_t seqLength = 0)
@@ -42,6 +43,31 @@ struct Query{
 
     Query() : queryId(0), classification(0), score(0), coverage(0), hammingDist(0), queryLength(0),
               queryLength2(0), kmerCnt(0), isClassified(false), newSpecies(false) {}
+};
+
+template<typename T>
+struct Buffer {
+    T *buffer;
+    size_t startIndexOfReserve;
+    size_t bufferSize;
+
+    explicit Buffer(size_t sizeOfBuffer=100) {
+        buffer = (T *) malloc(sizeof(T) * sizeOfBuffer);
+        bufferSize = sizeOfBuffer;
+        startIndexOfReserve = 0;
+    };
+
+    size_t reserveMemory(size_t numOfKmer) {
+        size_t offsetToWrite = __sync_fetch_and_add(&startIndexOfReserve, numOfKmer);
+        return offsetToWrite;
+    };
+
+    void reallocateMemory(size_t sizeOfBuffer) {
+        if (sizeOfBuffer > bufferSize) {
+            buffer = (T *) realloc(buffer, sizeof(T) * sizeOfBuffer);
+            bufferSize = sizeOfBuffer;
+        }
+    };
 };
 
 inline bool fileExist(const std::string& name) {
