+++ src/metabuli.cpp
@@ -12,7 +12,7 @@ const char* tool_name = "metabuli";
 const char* tool_introduction = "Metabuli is a taxonomical classifier that jointly analyzes amino acid and DNA sequences.";
 const char* main_author = "Jaebeom Kim <jbeom0731@gmail.com> ";
 const char* show_extended_help = "1";
-const char* show_bash_info = NULL;
+const char* show_bash_info = nullptr;
 bool hide_base_commands = true;
 extern const char* MMSEQS_CURRENT_INDEX_VERSION;
 const char* index_version_compatible = MMSEQS_CURRENT_INDEX_VERSION;
@@ -26,7 +26,7 @@ LocalParameters& localPar = LocalParameters::getLocalInstance();
 std::vector<Command> commands = {
         {"databases",            databases,            &localPar.databases,            COMMAND_DATABASE_CREATION,
                 "List and download databases",
-                NULL,
+                nullptr,
                 "Milot Mirdita <milot@mirdita.de>",
                 "<name> <o:sequenceDB> <tmpDir>",
                 CITATION_SPACEPHARER, {{"selection", 0, DbType::ZERO_OR_ALL, &DbValidator::empty },
@@ -34,7 +34,7 @@ std::vector<Command> commands = {
                                           {"tmpDir",     DbType::ACCESS_MODE_OUTPUT, DbType::NEED_DATA, &DbValidator::directory }}},
        {"build", build, &localPar.build, COMMAND_DATABASE_CREATION,
                 "Build database based on the list of FASTA files.",
-                NULL,
+               nullptr,
                 "Jaebeom Kim <jbeom0731@gmail.com>",
                 "<DB dir> <FASTA list> <Accesssion2taxid>",
                 CITATION_SPACEPHARER,
@@ -43,14 +43,14 @@ std::vector<Command> commands = {
                  {"Mapping file (accession to tax ID)", DbType::ACCESS_MODE_INPUT, DbType::NEED_DATA, &DbValidator::flatfile}}},
         {"database-report", databaseReport, &localPar.databaseReport, COMMAND_DATABASE_CREATION,
                 "It generates a report of taxa in a database.",
-                NULL,
+                nullptr,
                 "Jaebeom Kim <jbeom0731@gmail.com>",
                 "<i: DBDIR> ",
                 CITATION_SPACEPHARER,
                 {{"Directory where the DB will be generated", DbType::ACCESS_MODE_OUTPUT, DbType::NEED_DATA, &DbValidator::directory}}},
-       {"updateDB", build, &localPar.build, COMMAND_DATABASE_CREATION,
+       {"updateDB", build, &localPar.build, COMMAND_DB,
                "Update database based on the list of FASTA files.",
-               NULL,
+                nullptr,
                "Jaebeom Kim <jbeom0731@gmail.com>",
                "<DB dir> <FASTA list> <Accesssion2taxid>",
                CITATION_SPACEPHARER,
@@ -59,18 +59,25 @@ std::vector<Command> commands = {
                        {"Mapping file (accession to tax ID)", DbType::ACCESS_MODE_INPUT, DbType::NEED_DATA, &DbValidator::flatfile}}},
         {"classify", classify, &localPar.classify, COMMAND_TAXONOMY,
          "Assigning taxonomy label to query reads",
-         NULL,
+                nullptr,
          "Jaebeom Kim <jbeom0731@gmail.com>",
-         "<i:FASTA> <i:DB dir> <o:out dir> <job ID> ",
+         "<i:QUERY> <i:DB dir> <o:out dir> <job ID> ",
          CITATION_SPACEPHARER,
          {{"FASTA", DbType::ACCESS_MODE_INPUT, DbType::NEED_DATA | DbType::VARIADIC, &DbValidator::flatfile},
           {"DB dir", DbType::ACCESS_MODE_INPUT, DbType::NEED_DATA, &DbValidator::directory},
           {"out dir", DbType::ACCESS_MODE_OUTPUT, DbType::NEED_DATA, &DbValidator::directory},
           {"job ID", DbType::ACCESS_MODE_OUTPUT, DbType::NEED_DATA, &DbValidator::flatfile}}},
-
+        {"filter", classify, &localPar.classify, COMMAND_TAXONOMY,
+                "Filtering reads based on the classification result",
+                nullptr,
+                "Jaebeom Kim <jbeom0731@gmail.com>",
+                "<i:READ FILE> <i:FILTER DB>",
+                CITATION_SPACEPHARER,
+                {{"READ FILE", DbType::ACCESS_MODE_INPUT, DbType::NEED_DATA | DbType::VARIADIC, &DbValidator::flatfile},
+                        {"FILTER DB", DbType::ACCESS_MODE_INPUT, DbType::NEED_DATA, &DbValidator::directory}}},
         {"grade", grade, &localPar.grade, COMMAND_EXPERT,
                 "Grade the classification result (only for benchmarking)",
-                NULL,
+                nullptr,
                 "Jaebeom Kim <jbeom0731@gmail.com>",
                 "<i:Result list> <i:Answer sheet list> <i:Taxonomy Dir>",
                 CITATION_SPACEPHARER,
@@ -79,7 +86,7 @@ std::vector<Command> commands = {
                         {"taxonomy dir", DbType::ACCESS_MODE_INPUT, DbType::NEED_DATA | DbType::VARIADIC, &DbValidator::directory}}},
         {"seqHeader2TaxId", seqHeader2TaxId, &localPar.seqHeader2TaxId, COMMAND_EXPERT,
                 "It extracts k-mers from query sequences, and compares them to the target database",
-                NULL,
+                nullptr,
                 "Jaebeom Kim <jbeom0731@gmail.com>",
                 "<i:read-classification> <i:mapping>",
                 CITATION_SPACEPHARER,
@@ -88,7 +95,7 @@ std::vector<Command> commands = {
         {"add-to-library", addToLibrary, &localPar.addToLibrary, COMMAND_DATABASE_CREATION,
                     "It bins sequences into distinct files according to their species referring their accession number.\n "
                     "It requires a mapping file (accession to tax ID) and NCBI style tax dump files in a taxonomy directory.",
-                    NULL,
+                nullptr,
                     "Jaebeom Kim <jbeom0731@gmail.com>",
                     "<I: FASTA list> <I: accession2taxid> <I: DB DIR>",
                     CITATION_SPACEPHARER,
@@ -97,7 +104,7 @@ std::vector<Command> commands = {
                      {"DB directory", DbType::ACCESS_MODE_INPUT, DbType::NEED_DATA, &DbValidator::directory}}},
         {"apply-threshold", applyThreshold, &localPar.applyThreshold, COMMAND_MAIN,
                 "Assigning taxonomy label to query reads",
-                NULL,
+                nullptr,
                 "Jaebeom Kim <jbeom0731@gmail.com>",
                 "<i:Old readclassification> <o:OUT DIR> <o:JOB ID> <i: TAXONOMY DIR> ",
                 CITATION_SPACEPHARER,
@@ -107,7 +114,7 @@ std::vector<Command> commands = {
                         {"TAXONOMY DIR", DbType::ACCESS_MODE_OUTPUT, DbType::NEED_DATA, &DbValidator::directory}}},
         {"binning2report", binning2report, &localPar.binning2report, COMMAND_FORMAT_CONVERSION,
                 "It generates Kraken style report file from binning results",
-                NULL,
+                nullptr,
                 "Jaebeom Kim <jbeom0731@gmail.com>",
                 "<i:Binning Result> <o:OUT DIR> <o:JOB ID> <i: TAXONOMY DIR> ",
                 CITATION_SPACEPHARER,
@@ -117,7 +124,7 @@ std::vector<Command> commands = {
                         {"TAXONOMY DIR", DbType::ACCESS_MODE_OUTPUT, DbType::NEED_DATA, &DbValidator::directory}}},
         {"filter-by-genus", filterByGenus, &localPar.filterByGenus, COMMAND_EXPERT,
                     "It filters out reads classified as a specific genus",
-                    NULL,
+                nullptr,
                     "Jaebeom Kim <jbeom0731@gmail.com>",
                     "<i:Binning Result> <i:Genus list> <i: TAXONOMY DIR> ",
                     CITATION_SPACEPHARER,
