+++ src/workflow/CMakeLists.txt
@@ -3,4 +3,5 @@ set(workflow_source_files
         workflow/updateDB.cpp
         workflow/add_to_library.cpp
         workflow/build.cpp
+        workflow/filter.cpp
         PARENT_SCOPE)
